+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { ItemsComponent } from './components/watches/items.component';
+import { ItemsComponent } from './components/items/items.component';
 import { ItemDetailComponent } from './components/item-detail/item-detail.component';
 
 const routes: Routes = [
