+++ src/app/app.module.ts
@@ -1,21 +1,22 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';
+import { FormsModule } from '@angular/forms';
+import { CookieService } from 'ngx-cookie-service';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './components/header/header.component';
 import { PriceTagComponent } from './components/price-tag/price-tag.component';
-import { ItemsComponent } from './components/watches/items.component';
-import { ItemsGridViewComponent } from './components/watches/items-grid-view/items-grid-view.component';
-import { ItemsListViewComponent } from './components/watches/items-list-view/items-list-view.component';
-import { HttpClientModule } from '@angular/common/http';
+import { ItemsComponent } from './components/items/items.component';
+import { ItemsGridViewComponent } from './components/items/items-grid-view/items-grid-view.component';
+import { ItemsListViewComponent } from './components/items/items-list-view/items-list-view.component';
 import { ItemDetailComponent } from './components/item-detail/item-detail.component';
-import { FormsModule } from '@angular/forms';
 import { CartComponent } from './components/cart/cart.component';
 import { PaginationComponent } from './components/pagination/pagination.component';
 import { SidenavComponent } from './components/sidenav/sidenav.component';
 import { TooltipComponent } from './components/tooltip/tooltip.component';
-import { CookieService } from 'ngx-cookie-service';
+
 
 @NgModule({
     declarations: [
