+++ src/app/components/items/items-grid-view/items-grid-view.component.ts
