+++ src/app/components/items/items-list-view/items-list-view.component.ts
